Place tests into subfolder
