+++ src/dump/genial/tests/enum_multiple_variants.expected
