+++ src/dump/genial/tests/enum_single_variant_public.expected
