+++ src/dump/genial/tests/fn_single_line_body.expected
