+++ src/dump/genial/tests/structure_repr_c_public.expected
